Refactor map metric types and functions in PQDashboardController
