+++ src/PQDashboard/MapMetrics/MapMetric.cs
@@ -1,0 +1,66 @@
+﻿//******************************************************************************************************
+//  MapMetric.cs - Gbtc
+//
+//  Copyright © 2023, Grid Protection Alliance.  All Rights Reserved.
+//
+//  Licensed to the Grid Protection Alliance (GPA) under one or more contributor license agreements. See
+//  the NOTICE file distributed with this work for additional information regarding copyright ownership.
+//  The GPA licenses this file to you under the MIT License (MIT), the "License"; you may not use this
+//  file except in compliance with the License. You may obtain a copy of the License at:
+//
+//      http://opensource.org/licenses/MIT
+//
+//  Unless agreed to in writing, the subject software distributed under the License is distributed on an
+//  "AS-IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. Refer to the
+//  License for the specific language governing permissions and limitations.
+//
+//  Code Modification History:
+//  ----------------------------------------------------------------------------------------------------
+//  04/21/2023 - Stephen Wills
+//       Generated original version of source code.
+//
+//******************************************************************************************************
+
+namespace PQDashboard.MapMetrics
+{
+    public class MapMetric
+    {
+        public double Longitude;
+        public double Latitude;
+        public double? SizeValue;
+        public double? ColorValue;
+    }
+
+    public class MapMetricAggregate
+    {
+        public MapMetricAggregate(MapMetric mapMetric) { MapMetric = mapMetric; }
+
+        public MapMetric MapMetric { get; }
+        public double Maximum => _Maximum ?? 0.0D;
+        public double Minimum => _Minimum ?? 0.0D;
+        public double Average => Total / Count;
+
+        private double? _Maximum { get; set; }
+        private double? _Minimum { get; set; }
+        private double Total { get; set; }
+        private int Count { get; set; }
+
+        public void Aggregate(double value)
+        {
+            if (!_Maximum.HasValue || value > Maximum)
+                _Maximum = value;
+            if (!_Minimum.HasValue || value < Minimum)
+                _Minimum = value;
+            Total += value;
+            Count++;
+        }
+
+        public void Reset()
+        {
+            _Maximum = null;
+            _Minimum = null;
+            Total = 0.0D;
+            Count = 0;
+        }
+    }
+}
